fix/move/set up all db stuff to begin coding meat of human eval
