+++ IDEA/assignments/eval.json
@@ -1,5 +1,6 @@
 {
     "admin": {
+        "username": "admin",
         "password": "corbetsi"
     }
 }
